[restructuring] Refactoring plug-in names for 2.0.

+++ plugins/org.eclipse.recommenders.jayes/src/org/eclipse/recommenders/jayes/factor/arraywrapper/FloatArrayWrapper.java
@@ -85,7 +85,7 @@ public class FloatArrayWrapper implements IArrayWrapper {
 
     @Override
     public double getDouble(int index) {
-        return (double) array[index];
+        return array[index];
     }
 
     @Override
